Expanded tower
Made an upper level to the tower and started making an entrance to a bigger castle/tower. Also expanded the starting area slightly. Hoping to create a castle for the player to jump towards, but then they fall into a pit which sends them to hell, which is just the same area but destroyed and on fire