Added terrain tool and started ascent
Used the terrain tool to add a ground zero and put a tower near the bottom to be able to have the player teleport up it and into the clouds. Need to add clouds to the game to hide the ground below when the player looks down from the tower along with a death plane and respawn script for when the player fails the platforming.